requirements.txt
fastapi==0.95.1
googletrans==3.1.0a0
motor==2.3.0
pydantic==1.10.7
pymongo[srv]==3.13.0
python-dotenv==1.0.0
uvicorn==0.22.0